Extract common ru helpers to dedicated file
